feat(ui): landing page
The landing page consists of title, subtitle, and description about the content of the project.
Also a navigation button for the other pages which are the design 1 and 2 of the project.

feat(ui): design one

Feature-1 gallery consists of hero pages. Navigation button for home and next design. Also three sections of features that show unique design and layout.

feat(ui): design two

Feature-2 gallery consists of hero pages. Navigation button for home and next design. Also three sections of features that show unique design and layout

docs: update  in README

Updated the README file replace and input to TODO section.